app/requirements.txt
django==3.1.4
django-environ==0.4.5
psycopg2-binary==2.8.6
ebaysdk==2.2.0
django-extensions==3.1.0
celery==5.0.5
redis==3.5.3
django-storages[google]==1.10.1
Pillow==8.0.1
sentry-sdk==0.19.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
django-cors-headers==3.6.0